+++ wechat_django/models/__init__.py
@@ -4,13 +4,13 @@ from __future__ import unicode_literals
 
 from .constants import MsgLogFlag, MsgType
 
-from .apps import (MiniProgramApp, PublicApp, ServiceApp, SubscribeApp,
-                   WeChatApp)
+from .app import (MiniProgramApp, PublicApp, ServiceApp, SubscribeApp,
+                  WeChatApp)
 
 from .permission import permissions
 from .base import WeChatModel
 from .template import Template
-from .users import MiniProgramUser, PublicUser, UserTag, WeChatUser
+from .user import MiniProgramUser, PublicUser, Session, UserTag, WeChatUser
 from .material import Material
 from .article import Article
 from .messagehandler import MessageHandler
@@ -18,4 +18,3 @@ from .reply import Reply
 from .rule import Rule
 from .messagelog import MessageLog
 from .menu import Menu
-from .session import Session