+++ wechat_django/models/user/__init__.py
@@ -3,6 +3,7 @@
 from __future__ import unicode_literals
 
 from .base import WeChatUser
+from .session import Session
 from .miniprogram import MiniProgramUser
 from .public import PublicUser
 from .usertag import UserTag