+++ CHANGELOG.md
@@ -13,11 +13,13 @@ breaking changes以:exclamation:开头标注
 - [0.1.0](#010)
 
 ## 0.4.0
+本版本模型部分经过大量重构,虽通过单元测验,但并没有使用在生产环境当中,如有bug,请提issue
+
 * 重构app,user模型代码
 * [jssdk](docs/cookbook/web.md#jssdk)
 * [开放平台网站应用功能](docs/cookbook/web.md#开放平台网站应用)(待单元测试)
 * :exclamation: 调整微信支付服务商数据库字段,app新增服务商类型
-* 更新django-object-tool依赖,对于python3以上版本,将jsonfield依赖改为jsonfield2依赖,以兼容高版本django
+* 解决一些高版本python及django的兼容性问题,更新django-object-tool依赖,对于python3以上版本,将jsonfield依赖改为jsonfield2依赖
 
 ## 0.3.3
 * 给[首次关注](docs/cookbook/messages.md#首次订阅)用户加上属性